+++ src/display/surface.go
@@ -7,7 +7,7 @@ type Surface interface {
 	Fill()
 	FillPreserve()
 	SetLineWidth(width float64)
-	SetRgba(r, g, b, a float64)
+	SetRgba(r, g, b, a uint)
 	Stroke()
 
 	// Provides offset surface for nested components so that they can use
